move transition to disease.py
